@@ -12,60 +15,102 @@
     "dashboard",
     "extension"
   ],
-  "author": "Juha Mustonen <juha.p.mustonen@gmail.com>",
-  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/juhamust/mozaik-ext-slack.git"
+  },
   "dependencies": {
-    "babel-core": "6.7.6",
-    "babel-preset-es2015": "6.6.0",
-    "babel-preset-react": "6.5.0",
-    "babelify": "7.2.0",
-    "bluebird": "3.3.5",
+    "@babel/core": "^7.4.4",
+    "@mozaik/babel-preset": "^1.0.0-alpha.6",
+    "@mozaik/ui": "^2.0.0-alpha.15",
     "chalk": "1.1.3",
-    "classnames": "^2.2.5",
-    "convict": "^0.6.1",
-    "d3": "^4.1.0",
-    "d3-ease": "^1.0.0",
-    "dotenv": "^2.0.0",
-    "emojilib": "^2.0.2",
-    "lodash": "^3.10.1",
-    "pretty-print": "^1.0.0",
-    "react-classset": "0.0.2",
-    "react-mixin": "3.0.4",
-    "reflux": "0.4.1",
+    "color": "^3.1.2",
+    "convict": "^2.0.0",
+    "d3-ease": "^1.0.5",
+    "emojilib": "^2.4.0",
+    "moment": "^2.17.1",
+    "moment-duration-format": "1.3.0",
+    "nock": "^9.0.2",
+    "node-fetch": "^1.7.3",
+    "nyc": "^10.0.0",
+    "prettier": "^1.5.3",
+    "prop-types": "^15.7.2",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-feather": "^1.1.6",
+    "react-icons": "^2.2.5",
+    "react-test-renderer": "^15.5.0",
+    "reflux": "^6.4.1",
+    "request": "^2.79.0",
+    "request-promise-native": "^1.0.3",
+    "sha.js": "^2.4.10",
     "slack": "^7.3.0",
     "slack-remove-formatting": "^0.1.1",
-    "superagent": "1.8.3",
-    "superagent-bluebird-promise": "3.0.0",
-    "validator": "^3.30.0"
+    "styled-components": "^3.0.2",
+    "url-parse": "^1.2.0",
+    "util.promisify": "^1.0.0",
+    "vivus": "^0.4.3",
+    "whatwg-fetch": "^2.0.3"
   },
+  "main": "./lib/index.js",
+  "module": "es/index.js",
+  "jsnext:main": "es/index.js",
   "devDependencies": {
-    "ava": "^0.15.2",
-    "babel-eslint": "6.0.2",
-    "babel-register": "6.7.2",
-    "eslint": "^2.13.0",
-    "eslint-plugin-react": "^5.2.2",
-    "jsdom": "^9.2.1",
-    "proxyquire": "^1.7.9",
-    "touch": "0.0.3"
+    "@babel/preset-es2017": "^7.0.0-beta.53",
+    "ava": "^0.17.0",
+    "babel-cli": "^6.24.1",
+    "babel-eslint": "^7.2.3",
+    "babel-jest": "^20.0.3",
+    "babel-preset-es2015": "6.6.0",
+    "babel-preset-react": "6.5.0",
+    "coveralls": "^2.11.15",
+    "cross-env": "^5.0.1",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-15": "^1.0.5",
+    "enzyme-to-json": "^3.3.1",
+    "eslint": "^5.16.0",
+    "eslint-plugin-react": "^7.1.0",
+    "jest": "^20.0.4",
+    "jest-styled-components": "^4.11.0-0",
+    "lint-staged": "^4.0.2",
+    "watch": "^1.0.2"
   },
   "peerDependencies": {
+    "@mozaik/ui": "^2.0.0-alpha.15",
     "mozaik": ">=1.4.2",
-    "react": "^0.13.3"
+    "react": "^16.0.0"
+  },
+  "scripts": {
+    "lint": "eslint ./src/** ./test/**",
+    "lint:fix": "eslint --fix ./src/** ./test/**",
+    "test": "jest --verbose",
+    "test:cover": "jest --verbose --coverage",
+    "coverage": "cat ./coverage/lcov.info | coveralls",
+    "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir lib",
+    "build:es": "cross-env BABEL_ENV=es babel src --out-dir es",
+    "build": "npm run build:commonjs && npm run build:es",
+    "fmt": "prettier --print-width=100 --tab-width=2 --bracket-spacing --no-semi --trailing-comma es5 --single-quote --color --write \"{src,test}/**/*.js\" 'client.js'",
+    "fmt:check": "prettier --print-width=100 --tab-width=2 --bracket-spacing --no-semi --trailing-comma es5 --single-quote --list-different \"{src,test}/**/*.js\" 'client.js'",
+    "version": "echo ${npm_package_version}",
+    "watch": "watch 'npm run build' src"
+  },
+  "lint-staged": {
+    "*.js": [
+      "lint:fix",
+      "fmt",
+      "git add",
+      "jest --bail"
+    ]
+  },
+  "jest": {
+    "setupTestFrameworkScriptFile": "<rootDir>/tests/setupTests.js",
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ]
   },
   "browserify": {
     "transform": [
       "babelify"
     ]
-  },
-  "ava": {
-    "files": [
-      "tests/*.test.js"
-    ],
-    "tap": false,
-    "failFast": true,
-    "require": [
-      "babel-register"
-    ],
-    "babel": "inherit"
   }
 }
