@@ -1,2 +1,1 @@
-require('babel-register');
-module.exports = require('./src/client');
+module.exports = require('./lib/client');
