@@ -1,5 +1,6 @@
-import path from 'path';
-import convict from 'convict';
+const convict = require('convict');
+
+var path = require('path');
 
 const config = convict({
   slack: {
@@ -27,16 +28,24 @@ const config = convict({
       format: Boolean,
       env: 'SLACK_SHOW_IMAGES'
     },
-    // For testing/development
-    // NOTE: Use JSON in .env
-    // Example: SLACK_ECHO_MESSAGE = { "type": "message", "channel": "C02GVP9DZ", "user": "U02558FA2", "text": "Hello from Slack!", "ts": "1475003567.000006", "team": "T0254ARL8" }
     echoMessage: {
-      doc: 'Slack message to simulate',
-      default: {},
+      doc: 'A message to send regularly for testing',
+      default: null,
       format: Object,
       env: 'SLACK_ECHO_MESSAGE'
     }
+    // For testing/development
+    // NOTE: Use JSON in .env
+    // Example:
+    //   SLACK_ECHO_MESSAGE = {
+    //     "type":    "message",
+    //     "channel": "C02GVP9DZ",
+    //     "user":    "U02558FA2",
+    //     "text":    "Hello from Slack!",
+    //     "ts":      "1475003567.000006",
+    //     "team":    "T0254ARL8"
+    //   }
   }
 });
 
-export default config;
+exports.default = config;
